requirements.txt
httpx>=0.27.0
python-telegram-bot>=21.0
python-dotenv>=1.0
pydantic>=2.0
psutil>=5.9
aiohttp>=3.9
langchain-groq>=0.2
feedparser>=6.0
imaplib2>=3.6
gitpython>=3.1
pytest>=8.0
python-dateutil>=2.9
tweepy>=4.14.0
